Updated url_connector.c

+++ url_connector.c
@@ -353,7 +353,7 @@ int url_connection_send_request_ssl(int sd, char *write_buf, void *data, int (*c
 	//SSL READ
 	int count = 0;
 	char buf[1024];
-	long bytes;
+	int bytes;
 	while ((bytes = SSL_read(ssl, buf, sizeof(buf) -1)) >0 ) {
 		buf[bytes] = 0;
